less strictly follow defined themeCoords
as illustrated by "2024-05-19" and "2024-04-27", the order of characters
in a theme word isn't strictly followed by the official client. "EYES",
found in a 2x2 square in the board, can be selected in multiple orders,
and all are valid. To match here, we collect all possible positions for
a character in a word, then check against those.